+++ core/src/main/java/org/tron/easywork/handler/transfer/Trc10TransferHandler.java
@@ -7,7 +7,7 @@ import com.google.protobuf.InvalidProtocolBufferException;
 import org.tron.easywork.enums.TransferType;
 import org.tron.easywork.model.TransferInfo;
 import org.tron.easywork.model.Trc10TransferInfo;
-import org.tron.easywork.util.TransactionParser;
+import org.tron.easywork.util.TransactionUtil;
 import org.tron.trident.core.ApiWrapper;
 import org.tron.trident.proto.Chain;
 import org.tron.trident.proto.Contract;
@@ -61,7 +61,7 @@ public class Trc10TransferHandler extends BaseTransferHandler {
     @Override
     public Trc10TransferInfo getTransferInfo(GeneratedMessageV3 contract) {
         if (contract instanceof Contract.TransferAssetContract transferAssetContract) {
-            return TransactionParser.getTransferInfo(transferAssetContract);
+            return TransactionUtil.getTransferInfo(transferAssetContract);
         }
         throw new ClassCastException("不支持的类：" + contract.getClass().getName() + "。请提供 TransferAssetContract 类型对象!");
     }
