+++ core/src/test/java/org/tron/easywork/LocalTransferTest.java
@@ -10,9 +10,12 @@ import org.tron.easywork.demo.DemoLocalTransferHandler;
 import org.tron.easywork.handler.transfer.LocalTransferContext;
 import org.tron.easywork.handler.transfer.Trc20TransferHandler;
 import org.tron.easywork.handler.transfer.TrxTransferHandler;
-import org.tron.easywork.model.*;
+import org.tron.easywork.model.TransferInfo;
+import org.tron.easywork.model.Trc10TransferInfo;
+import org.tron.easywork.model.Trc20ContractInfo;
+import org.tron.easywork.model.Trc20TransferInfo;
 import org.tron.easywork.util.BlockParser;
-import org.tron.easywork.util.Trc20Utils;
+import org.tron.easywork.util.Trc20ContractUtil;
 import org.tron.trident.abi.FunctionEncoder;
 import org.tron.trident.abi.TypeReference;
 import org.tron.trident.abi.datatypes.Address;
@@ -55,7 +58,7 @@ public class LocalTransferTest extends BaseTest {
         // 实际转账金额
         BigDecimal realAmount = BigDecimal.valueOf(1);
         // 合约
-        Trc20ContractInfo trc20ContractInfo = Trc20Utils.readTrc20ContractInfo(testContractAddress, wrapper);
+        Trc20ContractInfo trc20ContractInfo = Trc20ContractUtil.readTrc20ContractInfo(testContractAddress, wrapper);
         // 系统转账金额
         BigDecimal transferAmount = trc20ContractInfo.getTransferAmount(realAmount);
         // 构造trc20交易
@@ -95,7 +98,7 @@ public class LocalTransferTest extends BaseTest {
         // 实际转账金额
         BigDecimal realAmount = BigDecimal.valueOf(1);
         // 合约
-        Trc20ContractInfo trc20ContractInfo = Trc20Utils.readTrc20ContractInfo(testContractAddress, wrapper);
+        Trc20ContractInfo trc20ContractInfo = Trc20ContractUtil.readTrc20ContractInfo(testContractAddress, wrapper);
         // 系统转账金额
         BigDecimal transferAmount = trc20ContractInfo.getTransferAmount(realAmount);
         // 构造trc20交易
@@ -233,7 +236,7 @@ public class LocalTransferTest extends BaseTest {
         // 转账金额
         BigDecimal realAmount = BigDecimal.valueOf(1);
         // 合约
-        Trc20ContractInfo trc20ContractInfo = Trc20Utils.readTrc20ContractInfo(testContractAddress, wrapper);
+        Trc20ContractInfo trc20ContractInfo = Trc20ContractUtil.readTrc20ContractInfo(testContractAddress, wrapper);
         // 转账金额
         BigDecimal transferAmount = trc20ContractInfo.getTransferAmount(realAmount);
         // 构造trc20 转账
