+++ core/src/test/java/org/tron/easywork/handler/collection/CollectionTest.java
@@ -5,7 +5,7 @@ import org.junit.Test;
 import org.tron.easywork.BaseTest;
 import org.tron.easywork.handler.transfer.Trc20TransferHandler;
 import org.tron.easywork.handler.transfer.TrxTransferHandler;
-import org.tron.easywork.util.Trc20Utils;
+import org.tron.easywork.util.Trc20ContractUtil;
 import org.tron.trident.core.exceptions.IllegalException;
 import org.tron.trident.proto.Chain;
 
@@ -45,7 +45,7 @@ public class CollectionTest extends BaseTest {
 
         // 归集配置
         FundCollectionConfig config = new FundCollectionConfig(
-                Trc20Utils.readTrc20ContractInfo("TFd1piJ8iXmJQicTicq4zChDSNSMLPFR4w", wrapper),
+                Trc20ContractUtil.readTrc20ContractInfo("TFd1piJ8iXmJQicTicq4zChDSNSMLPFR4w", wrapper),
                 "TBkg2JK18tY2YZArQgiS9DygzhVKptFirn",
                 "TBkg2JK18tY2YZArQgiS9DygzhVKptFirn",
                 "TP6QorvxAJ4bXg21LterCpGi5oZ2PxybCZ",
